# src/main/resources/static/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TGM Schulplaner API",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/todo_list/{id}": {
      "get": {
        "tags": [
          "todo"
        ],
        "summary": "Get todo list by id",
        "operationId": "getTodoList",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TodoList"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      },
      "put": {
        "tags": [
          "todo"
        ],
        "summary": "Get todo list by id",
        "operationId": "modifyTodoList",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyTodoList"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TodoList"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "todo"
        ],
        "summary": "Get todo list by id",
        "operationId": "deleteTodoList",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
    "/api/v1/user/todo_list": {
      "post": {
        "tags": [
          "user",
          "todo"
        ],
        "summary": "Create a new todo list that belong to the authenticated user",
        "operationId": "createTodoListForUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewTodoList"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TodoList"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
    "/api/v1/group/{gid}/todo_lists": {
      "get": {
        "tags": [
          "group",
          "todo"
        ],
        "summary": "Get all todo lists that belong to the group if the authenticated user has access",
        "operationId": "getTodoListsForGroup",
        "parameters": [
          {
            "name": "gid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoList"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
    "/api/v1/group/{gid}/todo_list": {
      "post": {
        "tags": [
          "group",
          "todo"
        ],
        "summary": "Create a new todo list that belongs to the group if the authenticated user has access",
        "operationId": "createTodoListForGroup",
        "parameters": [
          {
            "name": "gid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewTodoList"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TodoList"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
    "/api/v1/user/todo_lists": {
      "get": {
        "tags": [
          "user",
          "todo"
        ],
        "summary": "Get all todo lists that belong to the authenticated user",
        "operationId": "getTodoListsForUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoList"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
    "/api/v1/todo_list/{id}/item": {
      "post": {
        "tags": [
          "todo"
        ],
        "summary": "Create a new item for todo list by id",
        "operationId": "createTodoItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewTodoItem"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
    "/api/v1/todo_list/{id}/items": {
      "get": {
        "tags": [
          "todo"
        ],
        "summary": "Get all items from todo list by id",
        "operationId": "getTodoItems",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoItem"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
    "/api/v1/todo_list/{listId}/item/{id}/status": {
      "put": {
        "tags": [
          "todo"
        ],
        "summary": "Modify an item for todo list by id",
        "operationId": "modifyTodoItem",
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyTodoItemStatus"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
    "/api/v1/todo_list/{listId}/item/{id}": {
      "put": {
        "tags": [
          "todo"
        ],
        "summary": "Modify an item for todo list by id",
        "operationId": "modifyTodoItem_1",
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyTodoItem"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "todo"
        ],
        "summary": "Delete item for todo list by id",
        "operationId": "deleteTodoItem",
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TodoList": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "owner": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "NewTodoList": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          }
        }
      },
      "NewTodoItem": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "TodoItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "list": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "UNSTARTED",
              "IN_PROGRESS",
              "DONE",
              "FAILED"
            ]
          }
        }
      },
      "ModifyTodoList": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          }
        }
      },
      "ModifyTodoItemStatus": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "UNSTARTED",
              "IN_PROGRESS",
              "DONE",
              "FAILED"
            ]
          }
        }
      },
      "ModifyTodoItem": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer-key": {
        "type": "http",
        "name": "bearer-key",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "jwt"
      }
    }
  }
}